fix to bigInt

+++ resources/database/migrations/2016_08_12_061231_create_flash_sales_tbl.php
@@ -14,7 +14,7 @@ class CreateFlashSalesTbl extends Migration
     {
         Schema::create('flashsales', function(Blueprint $table){
             $table->increments('id');
-            $table->integer('user_id')->index()->unsigned();
+            $table->bigInteger('user_id')->index()->unsigned();
             $table->enum('type', ['single','group'])->default('single');
             $table->string('name');
             $table->text('description');
